Relocated friends-focused handlers to separate file from user handlers.
